Moved AuthErrorHandler to a separate top-level class
