+++ src/test/java/com/google/firebase/auth/FirebaseUserManagerTest.java
@@ -55,7 +55,6 @@ import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import org.junit.After;
